change macs_ble to 32 bit from 15 bit

+++ lib/libpax/globals.h
@@ -14,7 +14,7 @@
 #endif
 
 extern uint16_t volatile macs_wifi;
-extern uint16_t volatile macs_ble;
+extern uint32_t volatile macs_ble;
 
 extern TimerHandle_t WifiChanTimer;
 extern uint8_t volatile channel;  // wifi channel rotation counter
